feat: use extractor to place the flower inside the pot
