+++ assets/flower_pot_transformations.json
@@ -1,0 +1,39 @@
+{
+  "50": 383,
+  "237": 398,
+  "242": 403,
+  "247": 406,
+  "269": 761,
+  "51": 384,
+  "239": 400,
+  "230": 1102,
+  "55": 388,
+  "241": 402,
+  "238": 399,
+  "231": 1103,
+  "203": 391,
+  "244": 381,
+  "56": 389,
+  "229": 392,
+  "207": 407,
+  "205": 1085,
+  "240": 401,
+  "234": 395,
+  "243": 404,
+  "54": 387,
+  "251": 881,
+  "249": 879,
+  "235": 396,
+  "52": 385,
+  "53": 386,
+  "206": 1086,
+  "250": 880,
+  "248": 405,
+  "49": 382,
+  "236": 397,
+  "233": 394,
+  "328": 408,
+  "252": 882,
+  "57": 390,
+  "232": 393
+}
